Finished Initial SYMB/INFO Serialization Impl.
Finished initial implementations for all of the structures in the SYMB and INFO sections. It is now possible to deserialize and re-serialize these sections in their entirety without losing or altering any of the contained data. Tested and confirmed working with both "smashbros_sound.brsar" and "revo_kart.brsar".

Additionally cleared out FILE Section code in preparation for a proper rewrite.