Began Preparations to Remove "Contents" Byte Array
Started rewriting functions which currently reference the stored "contents" byteArray in the BRSAR struct. The goal is not to keep that stored at all, since we shouldn't need it after we parse everything into the structs.

The first of these are the functions for getting Group offsets and for summarizing the strings in the SYMB section (a test for this is in main.cpp).

Additionally, began const-qualifying some functions, may roll this back if I deem it unnecessary later.