Fixed Bug Preventing Setting 0 Bytes to Array
Fixed an issue with the bounds checking comparisons in setBytes which could cause it to reject requests to set 0 byte vectors. This didn't actually affect the array at all, since writing 0 bytes does nothing, but it does fix issues where the nextIndexOut value would be made invalid when it didn't need to be.