Finished Initial File Replace Impl, Removed Contents Array
Finished an initial implementation for replacing files in the FILE section. In doing so, implemented functions to handle updating file offsets and lengths in the INFO section (groupHeaders/Entries, fileHeaders) and in the FILE section (fileContents), as well as the BRSAR length count.

Additionally, rewrote the "exportSawnd" function to read from the parsed data in the brsar struct, finishing up the necessary preparations to completely eliminate the old "contents" byteArray stored.

Lastly, reworked "main.cpp" to use #define statements to trigger different tests. Also moved the "targetBrsarName" and "tempDumpBaseFolder" constants into "main.cpp".