Implemented .sawnd Parser and .sawnd Importer
Implemented "sawnd" data structure parser, as well as a function in the "brsar" struct for using the files they contain to overwrite existing file entries. Appropriate tests have been added to "main.cpp".

As a test of this utility, you'll find an alternative main function in main.cpp, which builds an executable that can be used as a replacement for the original "sawndz.exe" packaged with SuperSawndz.

Finally, fixed an minor issue which could cause brsarInfoGroupHeaders with no entries to not have their headerAddress and dataAddress fields updated by the "updateInfoSectionFileOffsets" function. Results of that function should now be identical to OG Sawndz.