Implemented Initial .wav/.dsp Conv. Automation
This should take care of the last remaining step in full .wav import and export support.

Additionally, as part of this implementation, functions have been added to the "rwsd" struct to streamline the import-export process a bit. Also added a credits notice to "lavaBRSARLib.h" rolled back version number to v0.9.3 to give room to finish implementing the other common file types.

Lastly, reworked "main.cpp" to use constexpr bools and normal "if" statements instead of pre-processor macros and #if directives. Helps to ensure that variables are scoped to the test block they're used in.