Prevented 0-Byte Requests in wavePacket Populate Func
Fixes a bug where, if the last wave in an RWSD got truncated, a request would be made to grab 0 bytes of padding from past the end of the provided data. This would fail, causing the entire parse to fail, which really isn't necessary since getting 0 bytes doesn't do anything anyway.

Note: I'm not fixing this with a change to lavaByteArray because I *always* want it to get upset when you ask for something past the array's endpoint, that should always fail.