Various Bug Fixes
Firstly, fixed a bug which could result in over-inflated WAVE sections when calling the "populate()" function on an "rwsd" struct more than once.

Second, fixed a bug which could sometimes prevent sounds from being properly exported to WAV using VGAudio. This arose from an asymmetry between the nibblesToSamples and samplesToNibbles functions; I'd mistakenly assumed that converting between the two would always yield identical results, when in actuality they do not. Fixed this by ensuring the nibbleCount value is set via the result of the samplesCount in exported DSP files, guaranteeing that things match up properly as far as VGAudio is concerned.

Lastly, fixed a minor issue in which VGAudio failing to convert a file wasn't being properly reflected in the return value of the exportWaveRawDataToWAV function.