Converted ByteArray From Template Class
ByteArray is no longer a templated class.
Previously, the only thing the templating actually controlled was the default endian type to use for its operations.
To move away from this, a new entry has been added to the endType enum which denotes a null value, which we use as the default function argument for all byte operations. When that null argument is detected, we then substitute in the defaultEndType member variable. This achieves the same effect ultimately, with the added benefit of being able to switch the default endian type at runtime.