WAVE Entries Can Now Accurately Calculate Their Size
This is important in that it should allow the pushBack/Front functions for the RWSD struct to properly support multi-channel entries.

This is also intended to be the start of a move to define methods and constants for interrogating the sizes of other structs as well, should be a good contribution to a general rewrite I've been planning.