Removed BRSAR Section Length Fields
The whole point of implementing the size functions was to get rid of these. With those functions in place, we can rely on calculating the values live, which is procedural and consistent, rather than trying to bookkeep them perfectly, which is more error prone.

The first payoff for this can be seen in the way the BRSAR header is now built. We no longer rely on maintained size or addresses, we can calculate everything now.

Also, implemented a size function for the BRSAR struct itself, forgot to do that initially.