InfoBankEntry Parent-Child System Impl.
This is something of a pilot for implementing this system for the rest of the structs, and initial tests suggest this will actually work! So core idea is that we store offsets from the parent for all children, and go "up the chain" so to speak to accurately calculate actual addresses as needed.