InfoPlayerEntry Parent-Child System Impl.
The same changes as last commit, except applied to the brsarInfoPlayerEntry struct.

Additionally, made a few other changes to how the system works. I've added a function argument to the infoSection size function, which makes it possible to easily get the size of the struct *up to a specific landmark*. We use this to more cleanly calculate the initial addresses for each offset update function and avoid duplicating so much code.

Lastly, the offset update functions are now void, as they have no relevant return value anymore since we don't care about the existence of a parent struct when we calculate this.