Finished File Section PCS Impl.
Additionally, implemented a caching system for section sizes. This came up because of how frequently the getAddress function is called in the File Section's functions; while I initially planned to just cache the returned address locally when the function was used, I figured doing it at the BRSAR level would simplify things, and extend those benefits to all the sections, not just the FILE section.

One change with this, though is that functions need to signal changes in their sizes to the BRSAR, which will then queue up a size update for the next time size is requested.