Removed Need For RefVecs in Info Section
The core 6 Reference Vectors are now built dynamically at export-time using the relevant entry vectors in the Info Section, which should guarantee they're always accurate, regardless of any reordering, additions, or removals of entries.

Note, though, that I still need to go in and give the same treatment to the refvecs in the sub-structures in the Info Section.

Also, slight readability improvements to the BrawlReference structs' size functions.