Moved Child-Struct Populate() Calls Into Parents
Previously, some child structs needed to be built in the InfoSection's Populate() function because we couldn't get the InfoSection's address to follow their brawlReferences to the relevant data. Now that we *can* do that, there's no reason to leave it externalized, so these have been moved in to their parent structs' populate() calls.