Removed Need For RefVecs in InfoGroupHeaders
These should now also be able to calculate their values as necessary, eliminating any need for book keeping on these values when shifting things around.

Also, the brawlReference that points to the list itself is now updated whenever we update the offsets of the groupEntries, so it should also be reliably accurate.