Some INFO Section Structs Now Stored As Unique_Ptr
This specifically was done to the structs which have child structs. Those child structs are holding pointers to their parents, and previously those parents were being stored directly in a vector, which meant that if the vector reallocated and moved everything the children would be left pointing to junk data. Storing unique_ptrs in these vectors instead means that the actual parent structs never move, which means their pointers should never be invalidated while their children are alive.