Implemented Clone File Test + Other Improvements
The Clone File Test is a pretty major milestone, it represents the first major payoff from implementing the parent-child system in the INFO section: we were able to successfully add a brand new file to the BRSAR, with minimal manual book keeping! The process is pretty much as simple as adding and populating the relevant entries and calling a couple update functions! A great deal of the tedium (managing BrawlRef addresses and BrawlRefVectors, mostly) is handled completely automatically, which is very very nice.

Additionally, I've ensured that brawlRefs in newly constructed structures get the right addressTypes when they have their addresses updated. I'm not sure what this actually would have done if left alone, but I expect the game would have crashed?

Second, the old individual update_____OffsetValues() functions in the INFO section have all been rolled into a single function, with a starting point parameter used to determine which landmark it'll start from when updating offset values. This came from a threefold realization:
- There is no situation in which you need to update one set of child structs' offsets and not for the sets after it
- Running the loop to update one set leaves you at the starting location for the next set
- The starting location of a set can be used to update the brawlRef to that set's brawlRefVec position, which needs to be done or else everything breaks
That all realized, rolling them all into a single function just works out to be the right solution. Guarantees that everything that needs to be updated is, and saves time over the old method since we're not needlessly recalculating the starting positions for each set of substructs.