Removed RWSD Header Struct + Improved RWSD Struct
The RWSD header struct was made largely obsolete by the introduction of the utilities put in place to manually calculate section sizes and offsets, and so it has been removed. The remaining relevant fields (endianType and versionNumber) have been rolled into the base RWSD struct.

Additionally, several utility functions for getting the sizes and offsets of the DATA and WAVE sections have been introduced, along with the same size caching system as is present in the BRSAR struct.