Implemented Function To Build RWSD From File
This was necessary for an addition to the CLI. Also does some basic validation of the file, checks for whether or not a LABL section is present (these exist in BrawlCrate exports, but aren't used for our purposes), and if so, ensures we ignore it. Works perfectly fine in my own testing.

Additionally, renamed a few of the hexTag enum entries to reflect the fact that they aren't actually RWSD exclusive, and can occur in different subfiles (eg. the DATA tag appears in all subfiles).