Added Func for Parsing ID List Documents
Will be used to facilitate the implementation of the "multi" variants of some of the existing CLI commands: rather than providing a single ID, you'll be able to pass in a path to a text document which specifies a list of IDs that the program will process as a batch.

Additionally, adjusted stringToNum function to be able to handle leading whitespaces on incoming strings.