Fixed Incorrect Data Addresses in RWSD Summaries
I'd forgotten to account for the fact that, when calculating the addresses for data entries using the offsets recorded in their brawlReference vector, you need to add them to 0x8 past the beginning of the Data section. The same is true of most things referred to using brawlReferences; though notably Wave entries don't use brawlReferences, and thus aren't subject to this change.